LPS-158464 Use data-engine-js-components-web type declarations
