+++ modules/apps/data-engine/data-engine-js-components-web/src/main/resources/META-INF/resources/js/index.d.ts
@@ -16,7 +16,11 @@ import {FieldType} from './utils/dataConverter';
 
 export {FieldFeedback} from './core/components/FieldFeedback';
 export {EVENT_TYPES as FORM_EVENT_TYPES} from './custom/form/eventTypes';
-export {getDDMFormFieldSettingsContext} from './utils/dataConverter';
+export {
+	getDDMFormFieldSettingsContext,
+	FieldType,
+	FieldTypeName,
+} from './utils/dataConverter';
 
 export function convertToFormData(body: unknown): unknown;
 
@@ -53,3 +57,16 @@ export function useForm(): ({
 }) => void;
 
 export function useFormState<T extends {[key: string]: unknown}>(): T;
+
+export const FormReport: React.FC<{
+	data?: string;
+	fields: unknown;
+	formReportRecordsFieldValuesURL: string;
+	portletNamespace: string;
+}>;
+
+export const FormView: React.FC;
+
+export const PartialResults: React.FC<{
+	reportDataURL: string;
+}>;
