+++ modules/apps/data-engine/data-engine-js-components-web/src/main/resources/META-INF/resources/ts_modules/frontend-js-web.d.ts
