+++ modules/apps/data-engine/data-engine-taglib/types/src/main/resources/META-INF/resources/data_layout_builder/js/plugins/fields-sidebar/components/SidebarFieldSettings.d.ts
@@ -12,10 +12,15 @@
  * details.
  */
 
+import {FieldTypeName} from 'data-engine-js-components-web';
 import React from 'react';
 import './SidebarFieldSettings.scss';
 declare const SidebarFieldSettings: React.FC<IProps>;
 export default SidebarFieldSettings;
+interface Field {
+	name: string;
+	type: FieldTypeName;
+}
 interface IProps {
 	field: Field;
 }
